Use 64-bit row counts in results and other changes

+++ h2/src/main/org/h2/command/CommandList.java
@@ -68,7 +68,7 @@ class CommandList extends Command {
     }
 
     @Override
-    public ResultInterface query(int maxrows) {
+    public ResultInterface query(long maxrows) {
         ResultInterface result = command.query(maxrows);
         executeRemaining();
         return result;
